removing old migrations and starting over
