+++ Cargo.toml
@@ -15,7 +15,7 @@ serde_derive = "1.0.80"
 simplelog = "0.5.3"
 strum = "0.11.0"
 strum_macros = "0.11.0"
-diesel = { version = "1.0.0", features = ["postgres"] }
+diesel = { version = "1.3.3", features = ["postgres"] }
 dotenv = "0.9.0"
 diesel-derive-enum = { version = "0.4", features = ["postgres"] }
